refactor(categories): extract form logic into useCategoryForm hook and add tests
